refactor(gift-cards): unify image lookup, add purpose to config
