+++ .env.example
@@ -19,7 +19,7 @@ LNURL_SERVER_ENCRYPTION_KEY="a8861cc3e5b3caf5573fbba2da2851a4e608a836eef10074be3
 # These mint + unit pairs cannot be added as an account. Set the unit to null to block the entire mint.
 # LNVoltz + USD: This mint has a bug that does not allow USD ecash to be melted. The current version is "Nutshell/0.18.1"
 VITE_CASHU_MINT_BLOCKLIST='[{"mintUrl":"https://mint.lnvoltz.com","unit":"usd"}]'
-VITE_GIFT_CARDS='[{"url":"https://blockandbean.agi.cash","name":"Block and Bean","currency":"BTC","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Block and Bean"],"nodePubkeys":[]}},{"url":"https://pubkey.agi.cash","name":"Pubkey","currency":"BTC","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Pubkey"],"nodePubkeys":[]}},{"url":"https://maple.agi.cash","name":"Maple","currency":"BTC","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Maple"],"nodePubkeys":[]}},{"url":"https://compass.agi.cash","name":"Compass Coffee","currency":"BTC","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Compass Coffee"],"nodePubkeys":[]}},{"url":"https://pinkowl.agi.cash","name":"Pink Owl Coffee","currency":"BTC","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Pink Owl Coffee"],"nodePubkeys":[]}},{"url":"https://shack.agi.cash","name":"The Shack","currency":"BTC","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["The Shack"],"nodePubkeys":[]}}]'
+VITE_GIFT_CARDS='[{"url":"https://blockandbean.agi.cash","name":"Block and Bean","currency":"BTC","purpose":"gift-card","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Block and Bean"],"nodePubkeys":[]}},{"url":"https://pubkey.agi.cash","name":"Pubkey","currency":"BTC","purpose":"gift-card","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Pubkey"],"nodePubkeys":[]}},{"url":"https://maple.agi.cash","name":"Maple","currency":"BTC","purpose":"gift-card","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Maple"],"nodePubkeys":[]}},{"url":"https://compass.agi.cash","name":"Compass Coffee","currency":"BTC","purpose":"gift-card","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Compass Coffee"],"nodePubkeys":[]}},{"url":"https://pinkowl.agi.cash","name":"Pink Owl Coffee","currency":"BTC","purpose":"gift-card","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["Pink Owl Coffee"],"nodePubkeys":[]}},{"url":"https://shack.agi.cash","name":"The Shack","currency":"BTC","purpose":"gift-card","isDiscoverable":true,"validPaymentDestinations":{"descriptions":["The Shack"],"nodePubkeys":[]}},{"url":"https://squarealphamay.agi.cash","name":"Square Alpha May","currency":"BTC","purpose":"offer","isDiscoverable":false,"validPaymentDestinations":{"descriptions":[],"nodePubkeys":[]}},{"url":"https://squaresantacruz.agi.cash","name":"Square Santa Cruz","currency":"BTC","purpose":"offer","isDiscoverable":false,"validPaymentDestinations":{"descriptions":[],"nodePubkeys":[]}}]'
 
 VITE_BREEZ_API_KEY='MIIBaTCCARugAwIBAgIHPxIwQzGEUzAFBgMrZXAwEDEOMAwGA1UEAxMFQnJlZXowHhcNMjYwNDA0MDg0MzIyWhcNMzYwNDAxMDg0MzIyWjApMRAwDgYDVQQKEwdBZ2ljYXNoMRUwEwYDVQQDEwxKb3NpcCBCb2pjaWMwKjAFBgMrZXADIQDQg/XL3yA8HKIgyimHU/Qbpxy0tvzris1fDUtEs6ldd6N7MHkwDgYDVR0PAQH/BAQDAgWgMAwGA1UdEwEB/wQCMAAwHQYDVR0OBBYEFNo5o+5ea0sNMlW/75VgGJCv2AcJMB8GA1UdIwQYMBaAFN6q1pJW843ndJIW/Ey2ILJrKJhrMBkGA1UdEQQSMBCBDmpvc2lwQGFnaS5jYXNoMAUGAytlcANBAK+gHYR+F8ssgm4LqWaom2FRIM9YjhCM6S1jh8Srh1APCf1TVGmPu4cRMmGWquTTe30dHObKfNvxI0SPFNhiSQA='
