+++ app/features/gift-cards/gift-card-config.ts
@@ -5,6 +5,7 @@ export const GiftCardConfigSchema = z.array(
     url: z.url(),
     name: z.string().min(1),
     currency: z.enum(['USD', 'BTC']),
+    purpose: z.enum(['gift-card', 'offer']),
     isDiscoverable: z.boolean(),
     addCardDisclaimer: z.string().optional(),
     validPaymentDestinations: z
@@ -22,3 +23,8 @@ export const JsonGiftCardConfigSchema = z
   .pipe(GiftCardConfigSchema);
 
 export type GiftCardConfig = z.infer<typeof GiftCardConfigSchema>[number];
+
+export type GiftCardInfo = GiftCardConfig & {
+  image: string;
+  ogImage: string | undefined;
+};
