+++ app/features/gift-cards/gift-card-details.tsx
@@ -10,8 +10,8 @@ import {
 import { Button } from '~/components/ui/button';
 import { getAccountBalance } from '~/features/accounts/account';
 import { useAccounts } from '~/features/accounts/account-hooks';
+import { getGiftCardImageByUrl } from '~/features/gift-cards/gift-card-images';
 import { GiftCardItem } from '~/features/gift-cards/gift-card-item';
-import { getGiftCardImageByUrl } from '~/features/gift-cards/use-discover-cards';
 import { MoneyWithConvertedAmount } from '~/features/shared/money-with-converted-amount';
 import { TransactionList } from '~/features/transactions/transaction-list';
 import { useBuildLinkWithSearchParams } from '~/hooks/use-search-params-link';
