+++ app/features/gift-cards/use-discover-cards.ts
@@ -1,64 +1,25 @@
 import { useMemo } from 'react';
-import blockAndBeanCard from '~/assets/gift-cards/blockandbean.agi.cash.webp';
-import compassCoffeeCard from '~/assets/gift-cards/compass.agi.cash.webp';
-import kissOfMatchaCard from '~/assets/gift-cards/kissofmatcha.agi.cash.webp';
-import mapleCard from '~/assets/gift-cards/maple.agi.cash.webp';
-import mariposaCard from '~/assets/gift-cards/mariposa.agi.cash.webp';
-import pinkOwlCoffeeCard from '~/assets/gift-cards/pinkowl.agi.cash.webp';
-import pubkeyCard from '~/assets/gift-cards/pubkey.agi.cash.webp';
-import theShackCard from '~/assets/gift-cards/shack.agi.cash.webp';
-import theEpicurianTraderCard from '~/assets/gift-cards/theepicureantrader.agi.cash.webp';
 import { useAccounts } from '../accounts/account-hooks';
 import {
-  type GiftCardConfig,
+  type GiftCardInfo,
   JsonGiftCardConfigSchema,
 } from './gift-card-config';
-
-export type GiftCardInfo = GiftCardConfig & {
-  image: string;
-  ogImage: string | undefined;
-};
-
-const GIFT_CARD_OG_IMAGES: Record<string, string> = {
-  'https://kissofmatcha.agi.cash': '/og/kissofmatcha.webp',
-  'https://mariposa.agi.cash': '/og/mariposa.webp',
-  'https://pinkowl.agi.cash': '/og/pinkowl.webp',
-  'https://pubkey.agi.cash': '/og/pubkey.webp',
-  'https://theepicureantrader.agi.cash': '/og/theepicureantrader.webp',
-};
-
-const GIFT_CARD_IMAGES: Record<string, string> = {
-  'https://blockandbean.agi.cash': blockAndBeanCard,
-  'https://pubkey.agi.cash': pubkeyCard,
-  'https://maple.agi.cash': mapleCard,
-  'https://mariposa.agi.cash': mariposaCard,
-  'https://compass.agi.cash': compassCoffeeCard,
-  'https://kissofmatcha.agi.cash': kissOfMatchaCard,
-  'https://pinkowl.agi.cash': pinkOwlCoffeeCard,
-  'https://shack.agi.cash': theShackCard,
-  'https://theepicureantrader.agi.cash': theEpicurianTraderCard,
-};
+import {
+  getGiftCardImageByUrl,
+  getGiftCardOgImageByUrl,
+} from './gift-card-images';
 
 function loadGiftCardsFromEnv(): GiftCardInfo[] {
   const raw = import.meta.env.VITE_GIFT_CARDS;
   if (!raw) return [];
-
-  // Validated at build time by vite.config.ts — safe to throw here.
   return JsonGiftCardConfigSchema.parse(raw).map((card) => ({
     ...card,
-    image: GIFT_CARD_IMAGES[card.url] ?? '',
-    ogImage: GIFT_CARD_OG_IMAGES[card.url],
+    image: getGiftCardImageByUrl(card.url) ?? '',
+    ogImage: getGiftCardOgImageByUrl(card.url),
   }));
 }
 
 export const GIFT_CARDS: GiftCardInfo[] = loadGiftCardsFromEnv();
-
-/**
- * Returns the gift card image for a given URL, if one exists.
- */
-export function getGiftCardImageByUrl(url: string): string | undefined {
-  return GIFT_CARDS.find((card) => card.url === url)?.image;
-}
 
 /**
  * Returns the gift card info for a given mint URL, if one exists.
