+++ app/features/send/find-matching-offer-or-gift-card-account.ts
@@ -3,7 +3,7 @@ import {
   type CashuAccount,
   getAccountBalance,
 } from '~/features/accounts/account';
-import type { GiftCardInfo } from '~/features/gift-cards/use-discover-cards';
+import type { GiftCardInfo } from '~/features/gift-cards/gift-card-config';
 import type { DecodedBolt11 } from '~/lib/bolt11';
 import { type Currency, Money } from '~/lib/money';
